Move comps to page.
